pyproject.toml
[build-system]
requires = [
  "setuptools>=63.0",
  "wheel"
]
build-backend = "setuptools.build_meta"

[project]
name             = "platypus-type-extensions"
version          = "0.1.0" 
description      = "Type and operator extensions for Platypus multiobjective optimization"   
authors = [
  {name="Shane Foster-Smith", email="shanefostersmith@users.noreply.github.com"},
]
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.26.0",
    "numba>=0.60.0",
    "scipy>=1.10.0",
    "platypus-opt>=1.4.1",
]
readme = "README.md"

[project.optional-dependencies]
distributions = [
    "pyomo>=6.8.2",
]
test = ["pytest>=7.0.0","pytest-mock>=3.10.0"]
full = [
  "numpy>=1.26.0",
  "numba>=0.60.0",
  "platypus-opt>=1.4.1",
  "pyomo>=6.8.2",
]


